data-pipeline/requirements.txt
# data-pipeline/requirements.txt
google-cloud-storage
google-cloud-discoveryengine
pypdf
python-magic
pytest
